refactor: break down TUI minimap.rs into focused modules
- Split 470-line minimap.rs into organized module hierarchy
- All files under 200 lines (largest: 134 lines)
- Group by function: state, builder, render, navigation, scroll
- Maintain all functionality and compilation
